fix integration test directory structure
