+++ .github/workflows/flaky-test-monitor.yml
@@ -21,7 +21,8 @@ jobs:
           - unit
           - unit-crypto
           - unit-integration
-          - integration-common
+          - integration-mvp
+          - integration-ghost
           - integration-network
           - integration-epochs
           - integration-access
