+++ .github/workflows/skipped-test-tracker-integration.yml
@@ -1,13 +1,13 @@
 name: Skipped Test Tracker Integration Test
 
 on:
-  # push:
-  #   branches:
-  #     - staging
-  #     - trying
-  # pull_request:
-  #   branches:
-  #     - master
+  push:
+    branches:
+      - staging
+      - trying
+  pull_request:
+    branches:
+      - master
 
 jobs:
   generate-skipped-test-summary:
@@ -19,7 +19,8 @@ jobs:
           - unit
           - unit-crypto
           - unit-integration
-          - integration-common
+          - integration-mvp
+          - integration-ghost
           - integration-network
           - integration-epochs
           - integration-access
