+++ integration/benchmark/tx_per_second_test.go
@@ -25,8 +25,8 @@ import (
 	"github.com/stretchr/testify/suite"
 	"google.golang.org/grpc"
 
-	"github.com/onflow/flow-go/integration/tests/common"
 	"github.com/onflow/flow-go/integration/tests/execution"
+	"github.com/onflow/flow-go/integration/tests/lib"
 	"github.com/onflow/flow-go/model/flow"
 	"github.com/onflow/flow-go/utils/unittest"
 )
@@ -223,7 +223,7 @@ func (gs *TransactionsPerSecondSuite) SetTokenAddresses() {
 func (gs *TransactionsPerSecondSuite) CreateAccountAndTransfer(keyIndex int) (flowsdk.Address, *flowsdk.AccountKey) {
 	ctx := context.Background()
 
-	myPrivateKey := common.RandomPrivateKey()
+	myPrivateKey := lib.RandomPrivateKey()
 	accountKey := flowsdk.NewAccountKey().
 		FromPrivateKey(myPrivateKey).
 		SetHashAlgo(crypto.SHA3_256).
