--- integration/tests/access/unstaked/main_test.go
@@ -1,12 +1,0 @@
-package unstaked
-
-import (
-	"testing"
-
-	"github.com/onflow/flow-go/integration/tests/access"
-	"github.com/stretchr/testify/suite"
-)
-
-func TestUnstakedAccessSuite(t *testing.T) {
-	suite.Run(t, new(access.UnstakedAccessSuite))
-}
