--- integration/tests/common/mvp/main_test.go
@@ -1,11 +1,0 @@
-package mvp
-
-import (
-	"testing"
-
-	"github.com/onflow/flow-go/integration/tests/common"
-)
-
-func TestMVP_Network(t *testing.T) {
-	common.MVP_Network(t)
-}
