--- integration/tests/common/mvp_bootstrap/main_test.go
@@ -1,11 +1,0 @@
-package mvp_bootstrap
-
-import (
-	"testing"
-
-	"github.com/onflow/flow-go/integration/tests/common"
-)
-
-func TestMVP_Bootstrap(t *testing.T) {
-	common.MVP_Bootstrap(t)
-}
