--- integration/tests/consensus/inclusion/main_test.go
@@ -1,12 +1,0 @@
-package inclusion
-
-import (
-	"testing"
-
-	"github.com/onflow/flow-go/integration/tests/consensus"
-	"github.com/stretchr/testify/suite"
-)
-
-func TestCollectionGuaranteeInclusion(t *testing.T) {
-	suite.Run(t, new(consensus.InclusionSuite))
-}
