--- integration/tests/epochs/epoch_join_and_leave_vn/main_test.go
@@ -1,15 +1,0 @@
-package epoch_join_and_leave_vn
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/suite"
-
-	"github.com/onflow/flow-go/integration/tests/epochs"
-	"github.com/onflow/flow-go/utils/unittest"
-)
-
-func TestEpoch(t *testing.T) {
-	unittest.SkipUnless(t, unittest.TEST_FLAKY, "epochs join/leave tests should be run on an machine with adequate resources")
-	suite.Run(t, new(epochs.EpochJoinAndLeaveVNSuite))
-}
