--- integration/tests/epochs/epoch_static_transition/main_test.go
@@ -1,13 +1,0 @@
-package epoch_static_transition
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/suite"
-
-	"github.com/onflow/flow-go/integration/tests/epochs"
-)
-
-func TestStaticTransition(t *testing.T) {
-	suite.Run(t, new(epochs.StaticEpochTransitionSuite))
-}
